wip: examples, last assertions
