+++ core/src/main/java/apoc/version/Version.java
@@ -9,7 +9,7 @@ import org.neo4j.procedure.UserFunction;
  */
 
 public class Version {
-// todo...
+
     @UserFunction("apoc.version")
     @Description("RETURN apoc.version() | return the current APOC installed version")
     public String version() {
