+++ docs/asciidoc/modules/ROOT/examples/generated-documentation/documentation.csv
@@ -184,13 +184,13 @@
 ¦procedure¦apoc.metrics.list¦apoc.metrics.list() :: (name :: STRING?, lastUpdated :: INTEGER?)¦apoc.metrics.list() - get a list of available metrics¦false¦
 ¦procedure¦apoc.metrics.storage¦apoc.metrics.storage(directorySetting :: STRING?) :: (setting :: STRING?, freeSpaceBytes :: INTEGER?, totalSpaceBytes :: INTEGER?, usableSpaceBytes :: INTEGER?, percentFree :: FLOAT?)¦apoc.metrics.storage(directorySetting) - retrieve storage metrics about the devices Neo4j uses for data storage. directorySetting may be any valid neo4j directory setting name, such as 'dbms.directories.data'.  If null is provided as a directorySetting, you will get back all available directory settings.  For a list of available directory settings, see the Neo4j operations manual reference on configuration settings.   Directory settings are **not** paths, they are a neo4j.conf setting key name¦false¦
 ¦procedure¦apoc.model.jdbc¦apoc.model.jdbc(jdbc :: STRING?, config = {} :: MAP?) :: (nodes :: LIST? OF NODE?, relationships :: LIST? OF RELATIONSHIP?)¦apoc.model.jdbc('key or url', {schema:'<schema>', write: <true/false>, filters: { tables:[], views: [], columns: []}) YIELD nodes, relationships - load schema from relational database¦false¦xref::database-integration/database-modeling.adoc
-¦procedure¦apoc.mongo.count¦apoc.mongo.count(uri :: STRING?, config = {} :: MAP?) :: (value :: INTEGER?)¦apoc.mongo.count(uri, $config) yield value - perform a find operation on mongodb collection¦true¦
-¦procedure¦apoc.mongo.delete¦apoc.mongo.delete(uri :: STRING?, query :: MAP?, config = {} :: MAP?) :: (value :: INTEGER?)¦apoc.mongo.delete(uri, query, $config) - delete the given documents from the mongodb collection and returns the number of affected documents¦true¦
-¦procedure¦apoc.mongo.find¦apoc.mongo.find(uri :: STRING?, config = {} :: MAP?) :: (value :: MAP?)¦apoc.mongo.find(uri, $config) yield value - perform a find operation on mongodb collection¦true¦
-¦procedure¦apoc.mongo.first¦apoc.mongo.first(uri :: STRING?, config = {} :: MAP?) :: (value :: MAP?)¦apoc.mongo.first(uri, $config) yield value - perform a find operation on mongodb collection¦true¦
-¦procedure¦apoc.mongo.get¦apoc.mongo.get(uri :: STRING?, config = {} :: MAP?) :: (value :: MAP?)¦apoc.mongo.get(uri, $config) yield value - perform a find operation on mongodb collection¦true¦
-¦procedure¦apoc.mongo.insert¦apoc.mongo.insert(uri :: STRING?, documents :: LIST? OF MAP?, config = {} :: MAP?) :: VOID¦apoc.mongo.insert(uri, documents, $config) yield value - inserts the given documents into the mongodb collection¦true¦
-¦procedure¦apoc.mongo.update¦apoc.mongo.update(uri :: STRING?, query :: MAP?, update :: MAP?, config = {} :: MAP?) :: (value :: INTEGER?)¦apoc.mongo.update(uri, query, update, $config) - updates the given documents from the mongodb collection and returns the number of affected documents¦true¦
+¦procedure¦apoc.mongo.count¦apoc.mongo.count(uri :: STRING?, config = {} :: MAP?) :: (value :: INTEGER?)¦apoc.mongo.count(uri, $config) yield value - perform a find operation on mongodb collection¦true¦xref::database-integration/mongo.adoc
+¦procedure¦apoc.mongo.delete¦apoc.mongo.delete(uri :: STRING?, query :: MAP?, config = {} :: MAP?) :: (value :: INTEGER?)¦apoc.mongo.delete(uri, query, $config) - delete the given documents from the mongodb collection and returns the number of affected documents¦true¦xref::database-integration/mongo.adoc
+¦procedure¦apoc.mongo.find¦apoc.mongo.find(uri :: STRING?, config = {} :: MAP?) :: (value :: MAP?)¦apoc.mongo.find(uri, $config) yield value - perform a find operation on mongodb collection¦true¦xref::database-integration/mongo.adoc
+¦procedure¦apoc.mongo.first¦apoc.mongo.first(uri :: STRING?, config = {} :: MAP?) :: (value :: MAP?)¦apoc.mongo.first(uri, $config) yield value - perform a find operation on mongodb collection¦true¦xref::database-integration/mongo.adoc
+¦procedure¦apoc.mongo.get¦apoc.mongo.get(uri :: STRING?, config = {} :: MAP?) :: (value :: MAP?)¦apoc.mongo.get(uri, $config) yield value - perform a find operation on mongodb collection¦true¦xref::database-integration/mongo.adoc
+¦procedure¦apoc.mongo.insert¦apoc.mongo.insert(uri :: STRING?, documents :: LIST? OF MAP?, config = {} :: MAP?) :: VOID¦apoc.mongo.insert(uri, documents, $config) yield value - inserts the given documents into the mongodb collection¦true¦xref::database-integration/mongo.adoc
+¦procedure¦apoc.mongo.update¦apoc.mongo.update(uri :: STRING?, query :: MAP?, update :: MAP?, config = {} :: MAP?) :: (value :: INTEGER?)¦apoc.mongo.update(uri, query, update, $config) - updates the given documents from the mongodb collection and returns the number of affected documents¦true¦xref::database-integration/mongo.adoc
 ¦procedure¦apoc.mongodb.count¦apoc.mongodb.count(host :: STRING?, db :: STRING?, collection :: STRING?, query :: MAP?) :: (value :: INTEGER?)¦apoc.mongodb.count(host-or-key,db,collection,query) yield value - perform a find operation on mongodb collection¦false¦xref::database-integration/mongodb.adoc
 ¦procedure¦apoc.mongodb.delete¦apoc.mongodb.delete(host :: STRING?, db :: STRING?, collection :: STRING?, query :: MAP?) :: (value :: INTEGER?)¦apoc.mongodb.delete(host-or-key,db,collection,query) - delete the given documents from the mongodb collection and returns the number of affected documents¦false¦xref::database-integration/mongodb.adoc
 ¦procedure¦apoc.mongodb.find¦apoc.mongodb.find(host :: STRING?, db :: STRING?, collection :: STRING?, query :: MAP?, project :: MAP?, sort :: MAP?, compatibleValues = false :: BOOLEAN?, skip = 0 :: INTEGER?, limit = 0 :: INTEGER?, extractReferences = false :: BOOLEAN?, objectIdAsMap = true :: BOOLEAN?) :: (value :: MAP?)¦apoc.mongodb.find(host-or-key,db,collection,query,projection,sort,[compatibleValues=false|true],skip-or-null,limit-or-null,[extractReferences=false|true],[objectIdAsMap=true|false]) yield value - perform a find,project,sort operation on mongodb collection¦false¦xref::database-integration/mongodb.adoc
