+++ docs/asciidoc/modules/ROOT/pages/database-integration/index.adoc
@@ -12,6 +12,7 @@ For more information on how to use these procedures, see:
 * xref::database-integration/load-jdbc.adoc[]
 * xref::database-integration/database-modeling.adoc[]
 * xref::database-integration/elasticsearch.adoc[]
+* xref::database-integration/mongo.adoc[]
 * xref::database-integration/mongodb.adoc[]
 * xref::database-integration/couchbase.adoc[]
 * xref::database-integration/bolt-neo4j.adoc[]
