+++ docs/asciidoc/modules/ROOT/pages/overview/apoc.mongo/apoc.mongo.count.adoc
@@ -32,3 +32,5 @@ apoc.mongo.count(uri :: STRING?, config = {} :: MAP?) :: (value :: INTEGER?)
 |value|INTEGER?
 |===
 
+xref::database-integration/mongo.adoc[More documentation of apoc.mongo.count,role=more information]
+
