+++ full/src/main/java/apoc/mongodb/MongoDBColl.java
@@ -36,7 +36,7 @@ import static java.lang.String.format;
  * @author mh
  * @since 30.06.16
  */
-class MongoDBColl implements MongoDB.Coll {
+class MongoDBColl implements MongoDBUtils.Coll {
 
     private static final ObjectMapper jsonMapper = new ObjectMapper().enable(DeserializationFeature.USE_LONG_FOR_INTS);
     public static final String ID = "_id";
