--- full/src/main/java/apoc/mongodb/MongoDbUtils.java
@@ -1,4 +1,0 @@
-package apoc.mongodb;
-
-public class MongoDbUtils {
-}
