+++ full/src/test/java/apoc/mongodb/MongoTest.java
@@ -6,6 +6,7 @@ import com.mongodb.MongoClient;
 import com.mongodb.MongoClientURI;
 import com.mongodb.client.MongoCollection;
 import com.mongodb.client.MongoDatabase;
+import org.apache.commons.collections.IteratorUtils;
 import org.apache.commons.lang3.time.DateUtils;
 import org.bson.BsonInt64;
 import org.bson.BsonRegularExpression;
@@ -52,35 +53,21 @@ import static org.junit.Assert.fail;
 
 public class MongoTest extends MongoTestBase {
 
-    @ClassRule
-    public static DbmsRule db = new ImpermanentDbmsRule();
+//    @ClassRule
+//    public static DbmsRule db = new ImpermanentDbmsRule();
 
-    private static MongoCollection<Document> testCollection;
-    private static MongoCollection<Document> productCollection;
-    private static MongoCollection<Document> personCollection;
     private static MongoCollection<Document> barCollection;
 
-    private static final Date currentTime = new Date();
-
-    private static final long longValue = 10_000L;
-
     private static String URI = null;
-
     private static List<String> BOUGHT_REFS;
+    private static final Set<String> SET_OBJECT_ID_MAP = Set.of("date", "machineIdentifier", "processIdentifier", "counter", "time", "timestamp", "timeSecond");
 
-
-    private static final long NUM_OF_RECORDS = 10_000L;
-
-    private static ObjectId nameAsObjectId = new ObjectId("507f191e810c19729de860ea");
     private static ObjectId fooAlAsObjectId = new ObjectId("77e193d7a9cc81b4027498b4");
     private static ObjectId fooJohnAsObjectId = new ObjectId("57e193d7a9cc81b4027499c4");
     private static ObjectId fooJackAsObjectId = new ObjectId("67e193d7a9cc81b4027518b4");
     private static ObjectId idAlAsObjectId = new ObjectId("97e193d7a9cc81b4027519b4");
     private static ObjectId idJackAsObjectId = new ObjectId("97e193d7a9cc81b4027518b4");
     private static ObjectId idJohnAsObjectId = new ObjectId("07e193d7a9cc81b4027488c4");
-    private static ObjectId idAsObjectId = new ObjectId();
-
-    private static final Set<String> SET_OBJECT_ID_MAP = Set.of("date", "machineIdentifier", "processIdentifier", "counter", "time", "timestamp", "timeSecond");
 
     @BeforeClass
     public static void setUp() throws Exception {
@@ -93,93 +80,58 @@ public class MongoTest extends MongoTestBase {
         final Integer port = mongo.getMappedPort(MONGO_DEFAULT_PORT);
         final String format = String.format("mongodb://admin:pass@%s:%s", host, port);
         MongoClientURI uri = new MongoClientURI(format);
-        MongoClient mongoClient = new MongoClient(uri);
+        try(MongoClient mongoClient = new MongoClient(uri)) {
+            URI = String.format("mongodb://admin:pass@%s:%s", mongo.getContainerIpAddress(), mongo.getMappedPort(MONGO_DEFAULT_PORT));
 
-        URI = String.format("mongodb://admin:pass@%s:%s", mongo.getContainerIpAddress(), mongo.getMappedPort(MONGO_DEFAULT_PORT));
+            fillDb(mongoClient);
 
-        // TODO - ESTRAPOLARE COMMON PART
-        TestUtil.registerProcedure(db, Mongo.class, Graphs.class);
+            personCollection.insertOne(new Document(map("_id", idAlAsObjectId,
+                    "name", "Al",
+                    "age", 25,
+                    "foo", fooAlAsObjectId,
+                    "date", new Date(1234567),
+                    "int64", new BsonInt64(29L),
+                    "double", 12.34,
+                    "expr", new BsonRegularExpression("foo*"),
+                    "binary", new Binary("fooBar".getBytes()),
+                    "bought", productReferences)));
+            personCollection.insertOne(new Document(map("_id", idJohnAsObjectId,
+                    "name", "John",
+                    "age", 45,
+                    "foo", fooJohnAsObjectId,
+                    "code", new Code("function() {}"),
+                    "sym", new Symbol("x"),
+                    "codeWithScope", new CodeWithScope("function() {}", new Document("k", "v")),
+                    "int64", new BsonInt64(29L),
+                    "time", new BsonTimestamp(123, 4),
+                    "minKey", new MinKey(),
+                    "maxKey", new MaxKey(),
+                    "expr", new BsonRegularExpression("foo*"),
+                    "binary", new Binary("fooBar".getBytes()),
+                    "bought", productReferences)));
+            personCollection.insertOne(new Document(map("_id", idJackAsObjectId,
+                    "name", "Jack",
+                    "age", 54,
+                    "foo", fooJackAsObjectId,
+                    "expr", new BsonRegularExpression("foo*"),
+                    "minKey", new MinKey(),
+                    "maxKey", new MaxKey(),
+                    "bought", productReferences)));
+            personCollection.insertOne(new Document(map("baz", "baa", "age", 200)));
+            personCollection.insertOne(new Document(map("baz", "baa", "age", 1)));
 
-
-
-        MongoDatabase database = mongoClient.getDatabase("test");
-        testCollection = database.getCollection("test");
-//        database.createCollection();
-        productCollection = database.getCollection("product");
-        personCollection = database.getCollection("person");
-        testCollection.deleteMany(new Document());
-        productCollection.deleteMany(new Document());
-        LongStream.range(0, NUM_OF_RECORDS)
-                .forEach(i -> testCollection.insertOne(new Document(map("name", "testDocument",
-                                    "date", currentTime, "longValue", longValue, "nullField", null))));
-
-        productCollection.insertOne(new Document(map("name", "My Awesome Product",
-                "price", 800,
-                "tags", Arrays.asList("Tech", "Mobile", "Phone", "iOS"))));
-        productCollection.insertOne(new Document(map("name", "My Awesome Product 2",
-                "price", 1200,
-                "tags", Arrays.asList("Tech", "Mobile", "Phone", "Android"))));
-        productReferences = StreamSupport.stream(productCollection.find().spliterator(), false)
-                .map(doc -> (ObjectId) doc.get("_id"))
-                .collect(Collectors.toList());
-        BOUGHT_REFS = productReferences.stream().map(ObjectId::toString).collect(Collectors.toList());
-
-        personCollection.insertOne(new Document(map("name", "Ajeje Brazorf",
-                "bought", productReferences,
-                "born", DateUtils.parseDate("11-10-1935", "dd-MM-yyyy"),
-                "coordinates", Arrays.asList(12.345, 67.890))));
-        personCollection.insertOne(new Document(map("name", nameAsObjectId,
-                "age", 40,
-                "bought", productReferences)));
-        personCollection.insertOne(new Document(map("_id", idAsObjectId,
-                "name", "Sherlock",
-                "age", 25,
-                "bought", productReferences)));
-        personCollection.insertOne(new Document(map("_id", idAlAsObjectId,
-                "name", "Al",
-                "age", 25,
-                "foo", fooAlAsObjectId,
-                "date", new Date(1234567),
-                "int64", new BsonInt64(29L),
-                "double", 12.34,
-                "expr", new BsonRegularExpression("foo*"),
-                "binary", new Binary("fooBar".getBytes()),
-                "bought", productReferences)));
-        personCollection.insertOne(new Document(map("_id", idJohnAsObjectId,
-                "name", "John",
-                "age", 45,
-                "foo", fooJohnAsObjectId,
-                "code", new Code("function() {}"),
-                "sym", new Symbol("x"),
-                "codeWithScope", new CodeWithScope("function() {}", new Document("k", "v")),
-                "int64", new BsonInt64(29L),
-                "time", new BsonTimestamp(123, 4),
-                "minKey", new MinKey(),
-                "maxKey", new MaxKey(),
-                "expr", new BsonRegularExpression("foo*"),
-                "binary", new Binary("fooBar".getBytes()),
-                "bought", productReferences)));
-        personCollection.insertOne(new Document(map("_id", idJackAsObjectId,
-                "name", "Jack",
-                "age", 54,
-                "foo", fooJackAsObjectId,
-                "expr", new BsonRegularExpression("foo*"),
-                "minKey", new MinKey(),
-                "maxKey", new MaxKey(),
-                "bought", productReferences)));
-
-        MongoDatabase databaseFoo = mongoClient.getDatabase("foo");
-        barCollection = databaseFoo.getCollection("bar");
-        barCollection.insertOne(new Document(map("name", nameAsObjectId, "age", 54)));
-
-
-        mongoClient.close();
+            BOUGHT_REFS = productReferences.stream().map(ObjectId::toString).collect(Collectors.toList());
+            MongoDatabase databaseFoo = mongoClient.getDatabase("foo");
+            barCollection = databaseFoo.getCollection("bar");
+            barCollection.insertOne(new Document(map("name", nameAsObjectId, "age", 54)));
+        }
     }
 
     private static Map<String, Object> getParams(String suffix) {
         return map("uri", URI + suffix);
     }
 
+    // - failing tests
     @Test(expected = QueryExecutionException.class)
     public void shouldFailIfUriHasNonRootUserNotAuthorizedInDb() {
         try {
@@ -203,14 +155,11 @@ public class MongoTest extends MongoTestBase {
                 });
     }
 
-    // todo - test connection without username e password --> che succede
     @Test(expected = QueryExecutionException.class)
     public void shouldFailIfUriHasNotCredentials() {
-
-        final String bytes = Base64.getEncoder().encodeToString("fooBar".getBytes());
         try {
-            testCall(db, "CALL apoc.mongo.get($uri, {query: {binary: {`$binary`: $bytes, `$subType`: '00'}}})",
-                    map("uri", String.format("mongodb://%s:%s/test.test?authSource=admin", mongo.getContainerIpAddress(), mongo.getMappedPort(MONGO_DEFAULT_PORT)), "bytes", bytes),
+            testCall(db, "CALL apoc.mongo.get($uri, {query: {foo: 'bar' }})",
+                    map("uri", String.format("mongodb://%s:%s/test.test?authSource=admin", mongo.getContainerIpAddress(), mongo.getMappedPort(MONGO_DEFAULT_PORT))),
                     row -> {});
         } catch (Exception e) {
             assertTrue(e.getMessage().contains("not authorized on test to execute command"));
@@ -242,10 +191,27 @@ public class MongoTest extends MongoTestBase {
         }
     }
 
+    @Test(expected = QueryExecutionException.class)
+    public void testFailsWithExtendedJsonFalseWithObjectIdParam() {
+        try {
+            testCall(db, "CALL apoc.mongo.first($uri,{query: {`_id`: {`$oid`: '97e193d7a9cc81b4027519b4'}}, useExtendedJson: false})",
+                    getParams("/test.person?authSource=admin"), r -> {});
+        } catch (QueryExecutionException e) {
+            assertTrue(e.getMessage().contains("unknown operator: $oid"));
+            throw e;
+        }
+    }
+
     @Test
     public void objectIdAsString() {
-        testResult(db, "CALL apoc.mongo.get($uri, {objectIdAsMap: false})", getParams("/test.person?authSource=admin"), res -> {
-            // todo
+        final String bytes = Base64.getEncoder().encodeToString("fooBar".getBytes());
+        testResult(db, "CALL apoc.mongo.get($uri, {query: {binary: {`$binary`: $bytes, `$subType`: '00'}}, objectIdAsMap: false})",
+                map("uri", URI + "/test.person?authSource=admin", "bytes", bytes),
+                res -> {
+            final ResourceIterator<Map<String, Object>> values = res.columnAs("value");
+            assertionsPersonAl(values.next(), false);
+            assertionsPersonJohn(values.next(), false);
+            assertFalse(values.hasNext());
         });
     }
 
@@ -254,8 +220,8 @@ public class MongoTest extends MongoTestBase {
         // todo - variabile statica con questo...
         testResult(db, "CALL apoc.mongo.get($uri, {query: {expr: {`$regex`: 'foo*', `$options`: ''} } , skip: 1})", getParams("/test.person?authSource=admin"), res -> {
             final ResourceIterator<Map<String, Object>> values = res.columnAs("value");
-            assertionsPersonJohn(values.next());
-            assertionsPersonJack(values.next());
+            assertionsPersonJohn(values.next(), true);
+            assertionsPersonJack(values.next(), true);
             assertFalse(values.hasNext());
         });
     }
@@ -264,8 +230,8 @@ public class MongoTest extends MongoTestBase {
     public void testWithLimit() {
         testResult(db, "CALL apoc.mongo.get($uri, {query: {expr: {`$regex`: 'foo*', `$options`: ''}} , limit: 2})", getParams("/test.person?authSource=admin"), res -> {
             final ResourceIterator<Map<String, Object>> values = res.columnAs("value");
-            assertionsPersonAl(values.next());
-            assertionsPersonJohn(values.next());
+            assertionsPersonAl(values.next(), true);
+            assertionsPersonJohn(values.next(), true);
             assertFalse(values.hasNext());
         });
     }
@@ -274,7 +240,7 @@ public class MongoTest extends MongoTestBase {
     public void testWithSkipAndLimit() {
         testResult(db, "CALL apoc.mongo.get($uri, {query: {expr: {`$regex`: 'foo*', `$options`: ''}} , skip: 1, limit: 1})", getParams("/test.person?authSource=admin"), res -> {
             final ResourceIterator<Map<String, Object>> value = res.columnAs("value");
-            assertionsPersonJohn(value.next());
+            assertionsPersonJohn(value.next(), true);
             assertFalse(value.hasNext());
         });
     }
@@ -283,9 +249,9 @@ public class MongoTest extends MongoTestBase {
     public void testFindSort() throws Exception {
         testResult(db, "CALL apoc.mongo.find($uri,{query: {expr: {`$regex`: 'foo*', `$options`: ''}}, sort: {name: -1}  })", getParams("/test.person?authSource=admin"), res -> {
             final ResourceIterator<Map<String, Object>> value = res.columnAs("value");
-            assertionsPersonJohn(value.next());
-            assertionsPersonJack(value.next());
-            assertionsPersonAl(value.next());
+            assertionsPersonJohn(value.next(), true);
+            assertionsPersonJack(value.next(), true);
+            assertionsPersonAl(value.next(), true);
             assertFalse(value.hasNext());
         });
     }
@@ -312,40 +278,11 @@ public class MongoTest extends MongoTestBase {
     }
 
     @Test
-    public void testFindManyConfigs() {
-        testResult(db, "CALL apoc.mongo.find($uri,{query: {expr: {`$regex`: 'foo*', `$options`: ''}}, project: {age: 1}  })", getParams("/test.person?authSource=admin"), res -> {
-            // TODO
+    public void testFindWithManyConfigs() {
+        testResult(db, "CALL apoc.mongo.find($uri,{query: {baz: 'baa'}, project: {age: 1}, sort: {name: -1}, {objectIdAsMap: false}, {skip: 1}, {limit: 2}})", getParams("/test.person?authSource=admin"), res -> {
+            // todo
+            IteratorUtils.toList(res.columnAs("value"));
         });
-    }
-
-    @Test
-    public void testGet()  {
-        testResult(db, "CALL apoc.mongo.get($uri)", getParams("/test.test?authSource=admin"), this::assertResult);
-    }
-
-    @Test
-    // todo - fail
-    public void testGetWithComplexTypes()  {
-        final String bytes = Base64.getEncoder().encodeToString("fooBar".getBytes());
-        testResult(db, "CALL apoc.mongo.get($uri, {query: {binary: {`$binary`: $bytes, `$subType`: '00'}}})",
-//        testResult(db, "CALL apoc.mongo.get($uri,{compatibleValues: true})",
-                map("uri", URI + "/test.person?authSource=admin", "bytes", bytes),
-                res -> {
-                    int count = 0;
-                    while (res.hasNext()) {
-                        ++count;
-                        Map<String, Object> r = res.next();
-                        Map doc = (Map) r.get("value");
-                        assertEquals(SET_OBJECT_ID_MAP, ((Map<String, Object>) doc.get("_id")).keySet());
-                        assertTrue(List.of("Al", "John").contains(doc.get("name")));
-                        assertTrue(List.of(25L, 45L).contains(doc.get("age")));
-                        final List<String> expectedObjIds = List.of("77e193d7a9cc81b4027498b4", "57e193d7a9cc81b4027499c4", "67e193d7a9cc81b4027518b4");
-                        assertTrue(expectedObjIds.contains(doc.get("foo")));
-                    }
-                    assertEquals(2, count);
-                });
-
-        // todo - get / find / first con altri tipi dato
     }
 
     @Test
@@ -354,7 +291,7 @@ public class MongoTest extends MongoTestBase {
             Map<String, Object> doc = (Map<String, Object>) r.get("value");
             assertEquals(Arrays.asList(12.345, 67.890), doc.get("coordinates"));
             assertEquals(LocalDateTime.of(1935,10,11, 0, 0), doc.get("born"));
-            assertEquals("Ajeje Brazorf", doc.get("name"));
+            assertEquals("Andrea Santurbano", doc.get("name"));
             assertEquals(BOUGHT_REFS, doc.get("bought"));
         });
     }
@@ -368,19 +305,7 @@ public class MongoTest extends MongoTestBase {
         });
     }
 
-    @Test(expected = QueryExecutionException.class)
-    public void testFailsWithExtendedJsonFalseWithObjectIdParam() {
-        try {
-            testCall(db, "CALL apoc.mongo.first($uri,{query: {`_id`: {`$oid`: '97e193d7a9cc81b4027519b4'}}, useExtendedJson: false})",
-                    getParams("/test.person?authSource=admin"), r -> {});
-        } catch (QueryExecutionException e) {
-            assertTrue(e.getMessage().contains("unknown operator: $oid"));
-            throw e;
-        }
-    }
-
     @Test
-    // todo - test con questo vero e compatibleValue falso e vedere che fa boom
     public void testFirstWithExtendedJson() {
         List<String> refsIds = productReferences.stream().map(ObjectId::toString).collect(Collectors.toList());
         testCall(db, "CALL apoc.mongo.first($uri,{query: {`_id`: {`$oid`: '97e193d7a9cc81b4027519b4'}}})",
@@ -395,60 +320,11 @@ public class MongoTest extends MongoTestBase {
                     assertEquals( 29L, doc.get("int64"));
                     assertEquals("77e193d7a9cc81b4027498b4", doc.get("foo") );
                     assertEquals(refsIds, doc.get("bought"));
-        });
-    }
-
-    @Test
-    public void testFind() throws Exception {
-        testResult(db, "CALL apoc.mongo.find($uri)", getParams("/test.test?authSource=admin"), this::assertResult);
-    }
-
-    @Test
-    public void testFindWithExtendedJson() throws Exception {
-        final String bytes = Base64.getEncoder().encodeToString("fooBar".getBytes());
-        testResult(db, "CALL apoc.mongo.find($uri,{query: {binary: {`$binary`: $bytes, `$subType`: '00'}}})",
-                map("uri", String.format("mongodb://user:secret@%s:%s/test.person?authSource=admin", mongo.getContainerIpAddress(), mongo.getMappedPort(MONGO_DEFAULT_PORT)), "bytes", bytes),
-
-                res -> {
-                    int count = 0;
-                    while (res.hasNext()) {
-                        ++count;
-                        Map<String, Object> r = res.next();
-                        Map doc = (Map) r.get("value");
-                        assertTrue(List.of("Al", "John").contains(doc.get("name")));
-                        assertEquals("foo*", doc.get("expr"));
-                        assertTrue(List.of(25L, 45L).contains(doc.get("age")));
-                        final List<String> expectedObjIds = List.of("97e193d7a9cc81b4027519b4", "07e193d7a9cc81b4027488c4");
-                        final Map<String, Object> id = (Map<String, Object>) doc.get("_id");
-                        assertEquals(SET_OBJECT_ID_MAP, id.keySet());
-                        final List<String> expectedObjIdsFooProp = List.of("77e193d7a9cc81b4027498b4", "57e193d7a9cc81b4027499c4");
-                        assertTrue(expectedObjIdsFooProp.contains(doc.get("foo")));
-                    }
-                    assertEquals(2, count);
                 });
     }
 
-    private void assertResult(Result res) {
-        assertResult(res, currentTime.toInstant().atZone(ZoneId.systemDefault()).toLocalDateTime());
-    }
-
-    private void assertResult(Result res, Object date) {
-        int count = 0;
-        while (res.hasNext()) {
-            ++count;
-            Map<String, Object> r = res.next();
-            Map doc = (Map) r.get("value");
-            assertNotNull(doc.get("_id"));
-            assertEquals("testDocument", doc.get("name"));
-            assertEquals(date, doc.get("date"));
-            assertEquals(longValue, doc.get("longValue"));
-        }
-        assertEquals(NUM_OF_RECORDS, count);
-    }
-
-
     @Test
-    public void testCountWithExtendedJson() throws Exception {
+    public void testCountWithComplexTypes() throws Exception {
         final String bytes = Base64.getEncoder().encodeToString("fooBar".getBytes());
         testCall(db, "CALL apoc.mongo.count($uri,{query: {binary: {`$binary`: 'Zm9vQmFy', `$subType`: '00'}, int64: {`$numberLong`: '29'}}})",
                 map("uri", URI + "/test.person?authSource=admin", "bytes", bytes),
@@ -456,13 +332,7 @@ public class MongoTest extends MongoTestBase {
     }
 
     @Test
-    public void testCountAll() throws Exception {
-        testCall(db, "CALL apoc.mongo.count($uri,{query: {name: 'testDocument'}})", getParams("/test.test?authSource=admin"),
-                r -> assertEquals(NUM_OF_RECORDS, r.get("value")));
-    }
-
-    @Test
-    public void testUpdateWithExtendedJson() throws Exception {
+    public void testUpdate() throws Exception {
         testCall(db, "CALL apoc.mongo.update($uri,{foo: {`$oid`: '57e193d7a9cc81b4027499c4'}},{`$set`:{code: {`$code`: 'void 0'}}})",
                 getParams("/test.person?authSource=admin"), r -> {
             long affected = (long) r.get("value");
@@ -478,7 +348,7 @@ public class MongoTest extends MongoTestBase {
     }
 
     @Test
-    public void testInsertWithExtendedJson() throws Exception {
+    public void testInsert() throws Exception {
         testResult(db, "CALL apoc.mongo.insert($uri,[{secondId: {`$oid` : '507f191e811c19729de860ea'}, baz: 1}, {secondId: {`$oid` : '507f191e821c19729de860ef'}, baz: 1}])",
                 map("uri", URI + "/test.person?authSource=admin"), (r) -> {
             assertFalse("should be empty", r.hasNext());
@@ -516,7 +386,7 @@ public class MongoTest extends MongoTestBase {
     }
 
     @Test
-    public void testGetInsertAndDeleteWithExtendedJson() throws Exception {
+    public void testGetInsertAndDeleteWithObjectId() throws Exception {
         final String suffix = "/test.person?authSource=admin";
         testResult(db, "CALL apoc.mongo.insert($uri,[{foo:'bar', myId: {`$oid` : '507f191e811c19729de960ea'}}])", getParams(suffix), (r) -> {
             assertFalse("should be empty", r.hasNext());
@@ -530,8 +400,25 @@ public class MongoTest extends MongoTestBase {
         });
     }
 
+    // -- tests similar to MongoDbTest.java to check consistency with old procedures
+    @Test
+    public void testFind() throws Exception {
+        testResult(db, "CALL apoc.mongo.find($uri)", getParams("/test.test?authSource=admin"), MongoTest::assertResult);
+    }
+
+    @Test
+    public void testGet()  {
+        testResult(db, "CALL apoc.mongo.get($uri)", getParams("/test.test?authSource=admin"), MongoTest::assertResult);
+    }
+
+    @Test
+    public void testCountAll() throws Exception {
+        testCall(db, "CALL apoc.mongo.count($uri,{query: {name: 'testDocument'}})", getParams("/test.test?authSource=admin"),
+                r -> assertEquals(NUM_OF_RECORDS, r.get("value")));
+    }
+
     @Test(expected = QueryExecutionException.class)
-    public void testInsertFailsDupKey() {
+    public void testInsertFailsWithDuplicateKey() {
         try {
             testResult(db, "CALL apoc.mongo.insert($uri,[{foo:'bar', _id: 1}, {foo:'bar', _id: 1}])", getParams("/test.test?authSource=admin"), (r) -> {
                 assertFalse("should be empty", r.hasNext());
@@ -549,26 +436,30 @@ public class MongoTest extends MongoTestBase {
                         "CALL apoc.graph.fromDocument(value, $fromDocConfig) YIELD graph AS g1 " +
                         "RETURN g1",
                 map("uri", URI + "/test.person?authSource=admin",
-                        "compatibleValues", true,
-                        "extractReferences", true,
-                        "fromDocConfig", map("write", true, "skipValidation", true, "mappings", map("$", "Person:Customer{!name,bought,coordinates,born}", "$.bought", "Product{!name,price,tags}"))),
+                        "fromDocConfig", map("write", true, "skipValidation", true, "mappings",
+                                map("$", "Person:Customer{!name,bought,coordinates,born}", "$.bought", "Product{!name,price,tags}"))),
                 r -> assertionsInsertDataWithFromDocument(date, r));
     }
 
-    private void assertionsPersonAl(Map<String, Object> second) {
-        assertEquals(25L, second.get("age"));
-        assertEquals(LocalDateTime.ofInstant(Instant.ofEpochMilli(1234567), ZoneId.systemDefault()), second.get("date"));
-        assertEquals(29L, second.get("int64"));
-        assertEquals(12.34, second.get("double"));
-        assertEquals(BOUGHT_REFS, second.get("bought"));
-        assertEquals(fooAlAsObjectId.toString(), second.get("foo"));
-        assertEquals("fooBar", new String((byte[]) second.get("binary")));
-        assertEquals("Al", second.get("name"));
-        assertEquals("foo*", second.get("expr"));
-        assertEquals(SET_OBJECT_ID_MAP, ((Map<String, Object>) second.get("_id")).keySet());
+
+    private void assertionsPersonAl(Map<String, Object> value, boolean idAsMap) {
+        assertEquals(25L, value.get("age"));
+        assertEquals(LocalDateTime.ofInstant(Instant.ofEpochMilli(1234567), ZoneId.systemDefault()), value.get("date"));
+        assertEquals(29L, value.get("int64"));
+        assertEquals(12.34, value.get("double"));
+        assertEquals(BOUGHT_REFS, value.get("bought"));
+        assertEquals(fooAlAsObjectId.toString(), value.get("foo"));
+        assertEquals("fooBar", new String((byte[]) value.get("binary")));
+        assertEquals("Al", value.get("name"));
+        assertEquals("foo*", value.get("expr"));
+        if (idAsMap) {
+            assertEquals(SET_OBJECT_ID_MAP, ((Map<String, Object>) value.get("_id")).keySet());
+        } else {
+            assertEquals(idAlAsObjectId.toString(), value.get("_id"));
+        }
     }
 
-    private void assertionsPersonJack(Map<String, Object> value) {
+    private void assertionsPersonJack(Map<String, Object> value, boolean idAsMap) {
         assertEquals(54L, value.get("age"));
         assertEquals("Jack", value.get("name"));
         assertEquals("MaxKey", value.get("maxKey"));
@@ -576,10 +467,14 @@ public class MongoTest extends MongoTestBase {
         assertEquals(BOUGHT_REFS, value.get("bought"));
         assertEquals("foo*", value.get("expr"));
         assertEquals(fooJackAsObjectId.toString(), value.get("foo"));
-        assertEquals(SET_OBJECT_ID_MAP, ((Map<String, Object>) value.get("_id")).keySet());
+        if (idAsMap) {
+            assertEquals(SET_OBJECT_ID_MAP, ((Map<String, Object>) value.get("_id")).keySet());
+        } else {
+            assertEquals(idJackAsObjectId.toString(), value.get("_id"));
+        }
     }
 
-    private void assertionsPersonJohn(Map<String, Object> value) {
+    private void assertionsPersonJohn(Map<String, Object> value, boolean idAsMap) {
         assertEquals(45L, value.get("age"));
         assertEquals("John", value.get("name"));
         assertEquals("function() {}", value.get("code"));
@@ -593,6 +488,10 @@ public class MongoTest extends MongoTestBase {
         assertEquals(123L, value.get("time"));
         assertEquals(fooJohnAsObjectId.toString(), value.get("foo"));
         assertEquals(29L, value.get("int64"));
-        assertEquals(SET_OBJECT_ID_MAP, ((Map<String, Object>) value.get("_id")).keySet());
+        if (idAsMap) {
+            assertEquals(SET_OBJECT_ID_MAP, ((Map<String, Object>) value.get("_id")).keySet());
+        } else {
+            assertEquals(idJohnAsObjectId.toString(), value.get("_id"));
+        }
     }
 }
